refactor: rename source files to match new verb names
- view.ts → get.ts, edit.ts → update.ts, bot/status.ts → bot/get.ts,
  sso/edit.ts → sso/update.ts
- Update all imports to reference new filenames
- Fix createSpinner → createSilentSpinner in bot/get.ts (pre-existing
  bug: was using nanospinner directly instead of the silent wrapper)
