+++ src/commands/app/bot/get.ts
@@ -1,6 +1,6 @@
 import { Command } from "commander";
 import pc from "picocolors";
-import { createSpinner } from "nanospinner";
+import { createSilentSpinner } from "../../../utils/spinner.js";
 import { getAccount, getTokenSilent, teamsDevPortalScopes } from "../../../auth/index.js";
 import { fetchApp, fetchAppDetailsV2, getBotLocation } from "../../../apps/index.js";
 import { pickApp } from "../../../utils/app-picker.js";
@@ -37,7 +37,7 @@ export const botGetCommand = new Command("get")
     }
 
     const botId = details.bots[0].botId;
-    const spinner = createSpinner("Checking bot location...").start();
+    const spinner = createSilentSpinner("Checking bot location...").start();
     const location = await getBotLocation(token, botId);
     spinner.stop();
