+++ src/commands/app/get.ts
