+++ src/commands/app/index.ts
@@ -2,8 +2,8 @@ import { Command } from "commander";
 import { select } from "@inquirer/prompts";
 import { appListCommand, runAppList } from "./list.js";
 import { appCreateCommand } from "./create.js";
-import { appGetCommand } from "./view.js";
-import { appUpdateCommand } from "./edit.js";
+import { appGetCommand } from "./get.js";
+import { appUpdateCommand } from "./update.js";
 import { appPackageCommand } from "./package/index.js";
 import { appManifestCommand } from "./manifest/index.js";
 import { authCommand } from "./auth/index.js";
